refactor: move Azure webhook logic to a dedicated module and clean up logging in student routes
